--- rapidfireai/evals/automl/__init__.py
@@ -1,21 +1,0 @@
-"""AutoML module for hyperparameter optimization."""
-
-from rapidfireai.evals.automl.base import AutoMLAlgorithm
-from rapidfireai.evals.automl.datatypes import List, Range
-from rapidfireai.evals.automl.grid_search import RFGridSearch
-from rapidfireai.evals.automl.model_config import RFLangChainRagSpec, RFPromptManager
-from rapidfireai.evals.automl.random_search import RFRandomSearch
-from rapidfireai.evals.automl.model_config import RFvLLMModelConfig, RFOpenAIAPIModelConfig, ModelConfig
-
-__all__ = [
-    "List",
-    "Range",
-    "RFGridSearch",
-    "RFRandomSearch",
-    "AutoMLAlgorithm",
-    "RFLangChainRagSpec",
-    "RFPromptManager",
-    "RFvLLMModelConfig",
-    "RFOpenAIAPIModelConfig",
-    "ModelConfig"
-]
