+++ rapidfireai/fit/utils/CLAUDE.md
@@ -337,7 +337,7 @@ if num_gpus == 0:
 
 **Usage**:
 ```python
-from rapidfireai.fit.utils.automl_utils import get_runs
+from rapidfireai.automl import get_runs
 
 # Handles both AutoML instances and plain dicts
 if isinstance(param_config, AutoMLAlgorithm):
